Makefile: All warnings are now errors. Removed unused functions, renamed gpuLoadVectors to gpuAllocateResults. Added casts where necessary to clear warnings. main: Changed parameter to gpuAllocateResults from vcnt to vcnt/pis, which is the correct count of rows. markkernel: Added offset parameter  to kernel launch to handle circuits with more than 1024 lines. Removed usage of merge LUTs, using straight boolean for now.

+++ sort.cc
@@ -19,7 +19,7 @@ int topologicalSort(NODE *graph, int ncount) {
 		if (graph[i].typ != 0)
 			ncnt++;
 	}
-	while (nodes.size() < ncnt) {
+	while (nodes.size() < (unsigned int)ncnt) {
 		for (int c = 0; c <= ncount; c++) {
 			if (graph[c].typ == 0) {
 				continue; // ignore blanks
